requirements.txt.txt
streamlit>=1.30.0
boto3>=1.34.0
requests>=2.31.0
pillow>=10.0.0
PyPDF2>=3.0.1
sounddevice>=0.4.6
soundfile>=0.12.1
python-dotenv